Refactor project structure and add acceptance tests for billing scenarios

+++ spec-tests/src/adapters.rs
@@ -181,5 +181,5 @@ fn ms_to_offset_datetime(epoch_ms: i64) -> Result<OffsetDateTime, ModelBError> {
     let seconds = epoch_ms.div_euclid(1_000);
     let millis = epoch_ms.rem_euclid(1_000);
     let base = OffsetDateTime::from_unix_timestamp(seconds)?;
-    Ok(base + Duration::milliseconds(millis as i64))
+    Ok(base + Duration::milliseconds(millis))
 }
